docs/requirements.txt
# Python dependencies
# run: pip install -r requirements.txt
alabaster==0.7.12
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
docutils==0.17.1
idna==3.2
imagesize==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
pyaml==21.8.3
Pygments==2.10.0
pyparsing==2.4.7
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==1.26.7
